refactor and cleanup carbon entity list tables
