+++ tornjak-frontend/src/tables/list-table.js
@@ -1,0 +1,64 @@
+import React from "react";
+import { DataTable } from "carbon-components-react";
+import ToolBar from './table-toolbar';
+import Head from './table-head';
+import Body from './table-body';
+const {
+    TableContainer,
+    Table,
+} = DataTable;
+
+class DataTableRender extends React.Component {
+    constructor(props) {
+        super(props);
+        this.state = {};
+    }
+
+    render() {
+        const { entityType, listTableData, headerData, deleteEntity, banEntity } = this.props;
+        return (
+            <DataTable
+                isSortable
+                rows={listTableData}
+                headers={headerData}
+                render={({
+                    rows,
+                    headers,
+                    getHeaderProps,
+                    getSelectionProps,
+                    onInputChange,
+                    getPaginationProps,
+                    getBatchActionProps,
+                    getTableContainerProps,
+                    selectedRows,
+                }) => (
+                    <TableContainer
+                        {...getTableContainerProps()}
+                    >
+                        <ToolBar
+                            onInputChange={onInputChange}
+                            getBatchActionProps={getBatchActionProps}
+                            deleteEntity={deleteEntity}
+                            banEntity={banEntity}
+                            selectedRows={selectedRows}
+                        />
+                        <Table size="short" useZebraStyles>
+                            <Head
+                                getHeaderProps={getHeaderProps}
+                                getSelectionProps={getSelectionProps}
+                                headers={headers}
+                            />
+                            <Body
+                                entityType={entityType}
+                                rows={rows}
+                                getSelectionProps={getSelectionProps}
+                            />
+                        </Table>
+                    </TableContainer>
+                )}
+            />
+        );
+    }
+}
+
+export default (DataTableRender)
